Add AmountWithUsdHuman component for consistent amount formatting
